show pool body on every route
