+++ src/components/swap/AdvancedSwapDetailsDropdown.tsx
@@ -12,10 +12,9 @@ const AdvancedDetailsFooter = styled.div<{ show: boolean }>`
   border-bottom-left-radius: 20px;
   border-bottom-right-radius: 20px;
   color: ${({ theme }) => theme.text2};
-  background-color: ${({ theme }) => theme.advancedBG};
+  background-color: ${({ theme, show }) => (show ? theme.advancedBG : 'clear')};
   z-index: -1;
-
-  transform: ${({ show }) => (show ? 'translateY(0%)' : 'translateY(-100%)')};
+  transform: ${({ show }) => (show ? 'translateY(0%) scaleY(0%)' : 'translateY(-100%) scaleY(0%)')};
   transition: transform 300ms ease-in-out;
 `
