+++ src/pages/Chef/index.tsx
@@ -7,7 +7,7 @@ import Question from '../../components/QuestionHelper'
 import Lock from '../../components/LockHelper'
 import { BalanceText } from '../../components/PositionCard'
 import { ChefPositionCard } from '../../components/ChefPositionCard'
-import { ExternalLink, StyledInternalLink, TYPE } from '../../theme'
+import { ExternalLink, TYPE } from '../../theme'
 import { Text } from 'rebass'
 import { LightCard } from '../../components/Card'
 import { RowBetween, RowFixed } from '../../components/Row'
@@ -17,7 +17,6 @@ import { AutoColumn } from '../../components/Column'
 import { useActiveWeb3React } from '../../hooks'
 import { useBlockNumber, useWalletModalToggle } from '../../state/application/hooks'
 import { useRewardToken, useUserInfoFarmablePools } from '../../data/Reserves'
-import AppBody from '../AppBody'
 import { Dots } from '../../components/swap/styleds'
 
 import { useBaocxBalance, useMasterChefContract } from '../../hooks/useContract'
@@ -109,147 +108,136 @@ export default function Chef() {
 
   return (
     <>
-      <AppBody>
-        <SwapPoolTabs active={'farm'} />
-        <AutoColumn gap="lg" justify="center">
-          <AutoColumn gap="12px" style={{ width: '100%' }}>
-            {chainId && masterChefContract && (
-              <RowBetween padding={'0 8px'}>
-                <ExternalLink id="link" href={getEtherscanLink(chainId, masterChefContract.address, 'address')}>
-                  BaoMasterFarmer Contract
-                  <TYPE.body color={theme.text3}>
-                    <b title={masterChefContract.address}>{shortenAddress(masterChefContract.address)} ↗</b>
-                  </TYPE.body>
-                </ExternalLink>
-                <AutoColumn gap="6px">
+      <SwapPoolTabs active={'farm'} />
+      <AutoColumn gap="lg" justify="center">
+        <AutoColumn gap="12px" style={{ width: '100%' }}>
+          {chainId && masterChefContract && (
+            <RowBetween padding={'0 8px'}>
+              <ExternalLink id="link" href={getEtherscanLink(chainId, masterChefContract.address, 'address')}>
+                BaoMasterFarmer Contract
+                <TYPE.body color={theme.text3}>
+                  <b title={masterChefContract.address}>{shortenAddress(masterChefContract.address)} ↗</b>
+                </TYPE.body>
+              </ExternalLink>
+              <AutoColumn gap="6px">
+                <RowBetween>
+                  <ButtonPrimary
+                    padding="0.5rem"
+                    onClick={() => handleHarvestAll()}
+                    disabled={attemptingHarvest || !account || v2IsLoading}
+                  >
+                    {attemptingHarvest ? (
+                      <span>
+                        <Dots>Harvesting</Dots>
+                        <IconWrapper pending={attemptingHarvest} success={!attemptingHarvest}>
+                          <Loader />
+                        </IconWrapper>
+                      </span>
+                    ) : (
+                      <span>
+                        <Text color={theme.text5} fontWeight={600}>
+                          Harvest All
+                        </Text>
+                        <BalanceText style={{ flexShrink: 0, textAlign: 'end' }} pr="0.5rem" fontWeight={800}>
+                          &nbsp;&nbsp;
+                          <UnlockIcon size="14px" /> {unlockedPending?.toFixed(0) || '-'}{' '}
+                          <span style={{ flexShrink: 1, fontSize: '8pt' }}>{rewardToken.symbol}</span>
+                          <br />
+                          + <LockIcon size="14px" /> {lockedPending?.toFixed(0) || '-'}{' '}
+                          <span style={{ flexShrink: 1, fontSize: '8pt' }}>{rewardToken.symbol}</span>
+                        </BalanceText>
+                      </span>
+                    )}
+                  </ButtonPrimary>
+                </RowBetween>
+                {baocxBalance?.greaterThan('0') ? (
                   <RowBetween>
-                    <ButtonPrimary
-                      padding="0.5rem"
-                      onClick={() => handleHarvestAll()}
-                      disabled={attemptingHarvest || !account || v2IsLoading}
-                    >
-                      {attemptingHarvest ? (
-                        <span>
-                          <Dots>Harvesting</Dots>
-                          <IconWrapper pending={attemptingHarvest} success={!attemptingHarvest}>
-                            <Loader />
-                          </IconWrapper>
-                        </span>
-                      ) : (
-                        <span>
-                          <Text color={theme.text5} fontWeight={600}>
-                            Harvest All
-                          </Text>
-                          <BalanceText style={{ flexShrink: 0, textAlign: 'end' }} pr="0.5rem" fontWeight={800}>
-                            &nbsp;&nbsp;
-                            <UnlockIcon size="14px" /> {unlockedPending?.toFixed(0) || '-'}{' '}
-                            <span style={{ flexShrink: 1, fontSize: '8pt' }}>{rewardToken.symbol}</span>
-                            <br />
-                            + <LockIcon size="14px" /> {lockedPending?.toFixed(0) || '-'}{' '}
-                            <span style={{ flexShrink: 1, fontSize: '8pt' }}>{rewardToken.symbol}</span>
-                          </BalanceText>
-                        </span>
-                      )}
-                    </ButtonPrimary>
+                    <ButtonSecondary padding="0.5rem" as={Link} to={`swap/${BAOCX.address}/${BAO.address}`}>
+                      <Text fontWeight={600}>Swap BAO.cx</Text>
+                      <ChevronRight />
+                    </ButtonSecondary>
                   </RowBetween>
-                  {baocxBalance?.greaterThan('0') ? (
-                    <RowBetween>
-                      <ButtonSecondary padding="0.5rem" as={Link} to={`swap/${BAOCX.address}/${BAO.address}`}>
-                        <Text fontWeight={600}>Swap BAO.cx</Text>
-                        <ChevronRight />
-                      </ButtonSecondary>
-                    </RowBetween>
-                  ) : (
-                    ''
-                  )}
-                </AutoColumn>
-              </RowBetween>
-            )}
-            <RowBetween padding={'0 8px'}>
-              <RowFixed>
-                <Text fontSize={16} fontWeight={500}>
-                  Locked {rewardToken.symbol}:
-                </Text>
-              </RowFixed>
-              <RowFixed>
-                <TYPE.body>
-                  <b>{lockedEarnedAmount.toFixed(2)}</b>
-                </TYPE.body>
-                <Question
-                  text={`Every time you Harvest or change your Stake, you instantly earn 5% of your pending rewards, and the remaining 95% will begin unlocking linearly at xDAI block ${unlockBlock}.`}
-                />
-                <Lock text={`Linear unlock begins in: ${(remainingBlocks / 12 / 60 / 24).toFixed(2)} days`} />
-              </RowFixed>
-            </RowBetween>
-
-            <RowBetween padding={'0 8px'}>
-              <Text color={theme.text1} fontWeight={500}>
-                Your Staked Liquidity Pools:
-              </Text>
-              <Question text="After you add liquidity to a pair, you are able to stake your position to earn BAOcx." />
-            </RowBetween>
-
-            {!account ? (
-              <LightCard padding="40px">
-                <ButtonLight onClick={toggleWalletModal}>Connect Wallet</ButtonLight>
-              </LightCard>
-            ) : v2IsLoading ? (
-              <LightCard padding="40px">
-                <TYPE.body color={theme.text3} textAlign="center">
-                  <Dots>Loading</Dots>
-                </TYPE.body>
-              </LightCard>
-            ) : userInfo?.length > 0 ? (
-              <>
-                {userInfo.map((farmablePool, i) => (
-                  <ChefPositionCard
-                    key={farmablePool.address}
-                    farmablePool={farmablePool}
-                    newRewardPerBlock={allNewRewardPerBlock[i]}
-                    baoPriceUsd={baoPriceUsd}
-                    unstakedLPAmount={tokenBalanceMap[farmablePool.address]}
-                    defaultShowMore={false}
-                  />
-                ))}
-              </>
-            ) : (
-              <LightCard padding="40px">
-                {noListSelected ? (
-                  <CurrencySearchModal
-                    isOpen={noListSelected}
-                    onCurrencySelect={() => {
-                      /* no-op */
-                    }}
-                    onDismiss={() => {
-                      /* no-op */
-                    }}
-                  />
                 ) : (
-                  <>
-                    <TYPE.body color={theme.text3} textAlign="center">
-                      No staked liquidity found.
-                    </TYPE.body>
-                    <ButtonPrimary id="join-pool-button" as={Link} style={{ marginTop: 16 }} to="/add/ETH">
-                      <Text fontWeight={500} fontSize={20}>
-                        Add Liquidity
-                      </Text>
-                    </ButtonPrimary>
-                  </>
+                  ''
                 )}
-              </LightCard>
-            )}
-
-            <div>
-              <Text textAlign="center" fontSize={14} style={{ padding: '.5rem 0 .5rem 0' }}>
-                {"Don't see a pool you joined?"}{' '}
-                <StyledInternalLink id="import-pool-link" to={'/find'}>
-                  {'Import it.'}
-                </StyledInternalLink>
+              </AutoColumn>
+            </RowBetween>
+          )}
+          <RowBetween padding={'0 8px'}>
+            <RowFixed>
+              <Text fontSize={16} fontWeight={500}>
+                Locked {rewardToken.symbol}:
               </Text>
-            </div>
-          </AutoColumn>
+            </RowFixed>
+            <RowFixed>
+              <TYPE.body>
+                <b>{lockedEarnedAmount.toFixed(2)}</b>
+              </TYPE.body>
+              <Question
+                text={`Every time you Harvest or change your Stake, you instantly earn 5% of your pending rewards, and the remaining 95% will begin unlocking linearly at xDAI block ${unlockBlock}.`}
+              />
+              <Lock text={`Linear unlock begins in: ${(remainingBlocks / 12 / 60 / 24).toFixed(2)} days`} />
+            </RowFixed>
+          </RowBetween>
+
+          <RowBetween padding={'0 8px'}>
+            <Text color={theme.text1} fontWeight={500}>
+              Your Staked Liquidity Pools:
+            </Text>
+            <Question text="After you add liquidity to a pair, you are able to stake your position to earn BAOcx." />
+          </RowBetween>
+
+          {!account ? (
+            <LightCard padding="40px">
+              <ButtonLight onClick={toggleWalletModal}>Connect Wallet</ButtonLight>
+            </LightCard>
+          ) : v2IsLoading ? (
+            <LightCard padding="40px">
+              <TYPE.body color={theme.text3} textAlign="center">
+                <Dots>Loading</Dots>
+              </TYPE.body>
+            </LightCard>
+          ) : userInfo?.length > 0 ? (
+            <>
+              {userInfo.map((farmablePool, i) => (
+                <ChefPositionCard
+                  key={farmablePool.address}
+                  farmablePool={farmablePool}
+                  newRewardPerBlock={allNewRewardPerBlock[i]}
+                  baoPriceUsd={baoPriceUsd}
+                  unstakedLPAmount={tokenBalanceMap[farmablePool.address]}
+                  defaultShowMore={false}
+                />
+              ))}
+            </>
+          ) : (
+            <LightCard padding="40px">
+              {noListSelected ? (
+                <CurrencySearchModal
+                  isOpen={noListSelected}
+                  onCurrencySelect={() => {
+                    /* no-op */
+                  }}
+                  onDismiss={() => {
+                    /* no-op */
+                  }}
+                />
+              ) : (
+                <>
+                  <TYPE.body color={theme.text3} textAlign="center">
+                    No staked liquidity found.
+                  </TYPE.body>
+                  <ButtonPrimary id="join-pool-button" as={Link} style={{ marginTop: 16 }} to="/add/ETH">
+                    <Text fontWeight={500} fontSize={20}>
+                      Add Liquidity
+                    </Text>
+                  </ButtonPrimary>
+                </>
+              )}
+            </LightCard>
+          )}
         </AutoColumn>
-      </AppBody>
+      </AutoColumn>
     </>
   )
 }
